+++ core/src/main/java/com/microsoft/applicationinsights/internal/perfcounter/JniPCConnector.java
@@ -79,11 +79,11 @@ public final class JniPCConnector {
             loadNativeLibrary();
         } catch (ThreadDeath td) {
             throw td;
+        } catch (JNIPerformanceCounterConnectorException e) {
+            logger.error("Error initializing JNI Performance Counter library. Windows performance counters will not be used.", e);
         } catch (Throwable e) {
             try {
-                logger.error(
-                    "Failed to load native dll, Windows performance counters will not be used. " +
-                "Please make sure that Visual C++ Redistributable is properly installed: {}.", e.toString());
+                logger.error("Unexpected error initializing JNI Performance Counter library. Windows performance counters will not be used", e);
 
                 return false;
             } catch (ThreadDeath td) {
@@ -154,6 +154,18 @@ public final class JniPCConnector {
         return getPerformanceCounterValue(name);
     }
 
+    /**
+     * Performance Counters identify a process by "Instance Name."
+     * This will be the executable name without the extension, e.g. a process running java.exe will have an instance name "java".
+     * If there are multiple instances of the same executable, an additional identifier is appended. By default this looks like "java#1", "java#2".
+     * For some reason, the instance name can change after the process starts with the default naming scheme.
+     *
+     * To workaround this, add a DWORD registry value named 'ProcessNameFormat' set to the value '2' to the key
+     * 'HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\PerfProc\Performance'.
+     * This changes the naming scheme from "java#2" to "java_PID" where PID is the current process id. It also makes the name constant for the life of the process.
+     *
+     * @throws NumberFormatException if pid cannot be parsed.
+     */
     private static void initNativeCode() {
         int processId = Integer.parseInt(SystemInformation.INSTANCE.getProcessId());
 
@@ -161,7 +173,7 @@ public final class JniPCConnector {
         if (StringUtils.isEmpty(currentInstanceName)) {
             logger.error("Failed to fetch current process instance name, process counters for for the process level will not be activated.");
         } else {
-            logger.trace("Java process name is set to '{}'", currentInstanceName);
+            logger.info("Java process instance name is set to '{}'", currentInstanceName);
         }
     }
 
@@ -177,21 +189,39 @@ public final class JniPCConnector {
      * @throws IOException If there are errors in opening/writing/reading/closing etc.
      *         Note that the method might throw RuntimeExceptions due to critical issues
      */
-    private static void loadNativeLibrary() throws IOException {
-        String model = System.getProperty("sun.arch.data.model");
-        String libraryToLoad = BITS_MODEL_64.equals(model) ? NATIVE_LIBRARY_64 : NATIVE_LIBRARY_32;
+    private static void loadNativeLibrary() throws JNIPerformanceCounterConnectorException {
+        final File dllOnDisk;
+        final String libraryToLoad;
+        try {
+            String model = System.getProperty("sun.arch.data.model");
+            libraryToLoad = BITS_MODEL_64.equals(model) ? NATIVE_LIBRARY_64 : NATIVE_LIBRARY_32;
 
-        File dllPath = buildDllLocalPath();
+            File dllPath = buildDllLocalPath();
 
-        File dllOnDisk = new File(dllPath, libraryToLoad);
+            dllOnDisk = new File(dllPath, libraryToLoad);
 
-        if (!dllOnDisk.exists()) {
-            extractToLocalFolder(dllOnDisk, libraryToLoad);
+            if (!dllOnDisk.exists()) {
+                extractToLocalFolder(dllOnDisk, libraryToLoad);
+            } else {
+                logger.trace("Found existing DLL: {}", dllOnDisk.getAbsolutePath());
+            }
+        } catch (Exception e) {
+            throw new JNIPerformanceCounterConnectorException("Error extracting DLL to disk", e);
         }
 
-        System.load(dllOnDisk.toString());
+        try {
+            System.load(dllOnDisk.toString());
+        } catch (Exception e) {
+            throw new JNIPerformanceCounterConnectorException("Error loading DLL. Please make sure that Visual C++ 2015 Redistributable is properly installed", e);
+        }
 
-        initNativeCode();
+        try {
+            initNativeCode();
+        } catch (NumberFormatException e) {
+            throw new JNIPerformanceCounterConnectorException("Could not parse PID as int", e);
+        } catch (Exception e) {
+            throw new JNIPerformanceCounterConnectorException("Unexpected error initializing performance counter DLL library", e);
+        }
 
         logger.trace("Successfully loaded library '{}'", libraryToLoad);
     }
@@ -246,5 +276,15 @@ public final class JniPCConnector {
         logger.trace("{} folder exists", dllPath.toString());
 
         return dllPath;
+    }
+
+    private static class JNIPerformanceCounterConnectorException extends Exception {
+        public JNIPerformanceCounterConnectorException(String s) {
+            super(s);
+        }
+
+        public JNIPerformanceCounterConnectorException(String s, Throwable throwable) {
+            super(s, throwable);
+        }
     }
 }